change Type of uniqueNumber to long in  getUniqueCustomerIdBasedOnNameAndFamily

+++ src/main/java/service/CustomerService.java
@@ -20,13 +20,13 @@ public class CustomerService {
     //Author : Mandana Khashayar
     //Return UniqueCustomerId based on name and family. It adds number between characters
     //in the way that the Id becomes unique
-    public String getUniqueCustomerIdBasedOnNameAndFamily(String name, String family, int uniqueNumber) {
+    public String getUniqueCustomerIdBasedOnNameAndFamily(String name, String family, long uniqueNumber) {
         var fullName=name.concat(family).toUpperCase(Locale.ROOT);
 
         List<Character> arrayList=  fullName.chars().mapToObj(c -> (char)c).collect(Collectors.toList());
         StringBuilder uniqueCustomerId
                 = new StringBuilder();
-        int uniqueIndex=fullName.length()+uniqueNumber-1;
+        long uniqueIndex=fullName.length()+uniqueNumber-1;
         for(int i=0;i<fullName.length();i++){
 
             uniqueCustomerId.append(arrayList.get(i)+String.valueOf(uniqueIndex-i));
